Develop backend

See merge request s09-final/S09P31A705!144

+++ backend/game-service/src/main/java/com/popplar/gameservice/dto/ConquerorInfoDto.java
@@ -17,10 +17,10 @@ public class ConquerorInfoDto {
     private Long hotPlaceId;
 
     @NotNull
-    private int points;
+    private double points;
 
     @Builder
-    public ConquerorInfoDto(Long id, Long memberId, Long hotPlaceId, int points) {
+    public ConquerorInfoDto(Long id, Long memberId, Long hotPlaceId, double points) {
         this.id = id;
         this.memberId = memberId;
         this.hotPlaceId = hotPlaceId;
